Added an encode() function
This functions builds upon tranlate and implements the math required to encode the ceaser cipher. Wrap around test have been conducted and were passed.

The translate function will still go into an infinite loop if passed a ascii code of something not between a:z. This will be fixed at a later date.